validity checking based on known parameters (#571)
* introduce check on parameters validity

* update __array__ validity requirements

* introduce is_unique and is_subrange_equal

* check if an array values are unique

* apply changes from master

* save to rebase

* fix flake reported error

* disable failing wip test

* scale down unique api, add kernels and tests for indexedarray

* fix python code for generated tests

* cleanup kernels

* remove spurious includes

* final cleanup

* cleanup the rest of the warnings

* Centralized parameter-checking in 'Content::validityerror_parameters'.

* Black and flake8.

+++ src/libawkward/Identities.cpp
@@ -232,7 +232,7 @@ namespace awkward {
   const IdentitiesPtr
   IdentitiesOf<T>::deep_copy() const {
     std::shared_ptr<T> ptr = kernel::malloc<T>(kernel::lib::cpu,   // DERIVE
-                                               length_*sizeof(T));
+                                               length_*(int64_t)sizeof(T));
     if (length_ != 0) {
       // DERIVE Yikes! It's a memcpy!
       memcpy(ptr.get(),
